Implements a generic class structure for query model classes
There is no shady casting required anymore when accessing shared methods implement in abstract queries classes anymore since we know now the type of the concrete query at hand.
On the other hand, it is now required to explicitly state the query class when defining a new query class since Java does not allow to reference implicitly a class' own type in generics.