Changes return type of getDistVarsWithVarType to return a copy of the internal map
Previously, accidentially only an undmofiable view on the map was returned but this did not mean that the values (=sets of dist. vars) were unmodifiable as well - they were thus changed and side effects occurred (e.g. deleting of dist. vars in a copy resulted in deleting dist. vars in the original query as wwell).