Removes fully Boolean CNCQ engine
We now only rely on the semi-Boolean CNCQ engine which does everythting the Boolean engine did before. This reduces duplicate code.
Also, we now have a non-Boolean CNCQ engine that iterates through all bindings and calls the semi-Boolean engine to fetch the remaining bindings (for the UCQ part), but this is to be tested.