Hotfixes a bug where object properties could not be deleted due to the core reasoner deleting the subject before
This is only a hotfix as a proper cleanup should be conducted even in this case. Or better, the reasoner shall not remove such individuals.
Right now it does because the branch index of the individual is higher than the branch index of the reasoning state.
Manipulating the individual dependency set does not help. Maybe a correct parent has to be set.

Another approach would be to not rely on individuals at all but rather roll the positive part of the CNCQ query up and use assert C subseteq T in the TBox.