Refactors query engines to point to the abox they are considering
Previously, only the abox was taken from the query's KB. But this may be the wrong abox in case of nesting query engines that at
some point manipulate the ABoxes. Then, we the called query engine needs to operate on the modified ABox and not on the original
ABox of the query.